Modified the 10-binary_tree_depth.c file and added a comment.

+++ 10-binary_tree_depth.c
@@ -8,15 +8,17 @@
 */
 size_t binary_tree_depth(const binary_tree_t *tree)
 {
-	int dep = 0;
+	size_t dep = 0; /* Variable to store the depth */
 
-	if (tree == NULL)
+	if (tree == NULL) /* If tree is NULL, return 0 */
 		return (0);
-	while (tree->parent != NULL)
+
+	/*Traverse up the tree until reaching the root*/
+		while (tree->parent != NULL)
 	{
-		dep++;
-		tree = tree->parent;
+		dep++; /* Increment depth for each parent */
+		tree = tree->parent; /* Move to the parent node */
 	}
 
-	return (dep);
+	return (dep); /* Return the calculated depth */
 }
